09/07/2025 feat:Cambia a la version 8.1 del backend

+++ dspace-api/src/main/java/org/dspace/orcid/service/impl/OrcidSynchronizationServiceImpl.java
@@ -19,7 +19,6 @@ import static org.dspace.profile.OrcidEntitySyncPreference.DISABLED;
 
 import java.sql.SQLException;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -32,24 +31,25 @@ import org.dspace.content.MetadataValue;
 import org.dspace.content.service.ItemService;
 import org.dspace.core.Context;
 import org.dspace.discovery.DiscoverQuery;
-import org.dspace.discovery.DiscoverResultItemIterator;
+import org.dspace.discovery.SearchService;
+import org.dspace.discovery.SearchServiceException;
 import org.dspace.discovery.indexobject.IndexableItem;
 import org.dspace.eperson.EPerson;
 import org.dspace.eperson.service.EPersonService;
-import org.dspace.orcid.OrcidQueue;
 import org.dspace.orcid.OrcidToken;
+import org.dspace.orcid.client.OrcidClient;
 import org.dspace.orcid.model.OrcidEntityType;
 import org.dspace.orcid.model.OrcidTokenResponseDTO;
-import org.dspace.orcid.service.OrcidQueueService;
 import org.dspace.orcid.service.OrcidSynchronizationService;
 import org.dspace.orcid.service.OrcidTokenService;
-import org.dspace.orcid.service.OrcidWebhookService;
 import org.dspace.profile.OrcidEntitySyncPreference;
 import org.dspace.profile.OrcidProfileDisconnectionMode;
 import org.dspace.profile.OrcidProfileSyncPreference;
 import org.dspace.profile.OrcidSynchronizationMode;
 import org.dspace.profile.service.ResearcherProfileService;
 import org.dspace.services.ConfigurationService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -60,26 +60,27 @@ import org.springframework.beans.factory.annotation.Autowired;
  */
 public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationService {
 
+    private static final Logger log = LoggerFactory.getLogger(OrcidSynchronizationServiceImpl.class);
     @Autowired
     private ItemService itemService;
-
-    @Autowired
-    private OrcidQueueService orcidQueueService;
 
     @Autowired
     private ConfigurationService configurationService;
 
     @Autowired
-    private OrcidWebhookService orcidWebhookService;
+    private EPersonService ePersonService;
 
     @Autowired
-    private EPersonService ePersonService;
+    private SearchService searchService;
 
     @Autowired
     private OrcidTokenService orcidTokenService;
 
     @Autowired
     private ResearcherProfileService researcherProfileService;
+
+    @Autowired
+    private OrcidClient orcidClient;
 
     @Override
     public void linkProfile(Context context, Item profile, OrcidTokenResponseDTO token) throws SQLException {
@@ -120,27 +121,31 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
     @Override
     public void unlinkProfile(Context context, Item profile) throws SQLException {
 
-        if (orcidWebhookService.isProfileRegistered(profile)) {
-            orcidWebhookService.unregister(context, profile);
-        }
+        clearOrcidProfileMetadata(context, profile);
 
-        itemService.clearMetadata(context, profile, "person", "identifier", "orcid", Item.ANY);
-        itemService.clearMetadata(context, profile, "dspace", "orcid", "scope", Item.ANY);
-        itemService.clearMetadata(context, profile, "dspace", "orcid", "authenticated", Item.ANY);
+        clearSynchronizationSettings(context, profile);
 
-        if (!configurationService.getBooleanProperty("orcid.disconnection.remain-sync", false)) {
-            clearSynchronizationSettings(context, profile);
-        }
-
-        orcidTokenService.deleteByProfileItem(context, profile);
+        clearOrcidToken(context, profile);
 
         updateItem(context, profile);
 
-        List<OrcidQueue> queueRecords = orcidQueueService.findByProfileItemId(context, profile.getID());
-        for (OrcidQueue queueRecord : queueRecords) {
-            orcidQueueService.delete(context, queueRecord);
+    }
+
+    private void clearOrcidToken(Context context, Item profile) {
+        OrcidToken profileToken = orcidTokenService.findByProfileItem(context, profile);
+        if (profileToken == null) {
+            log.warn("Cannot find any token related to the user profile: {}", profile.getID());
+            return;
         }
 
+        orcidTokenService.deleteByProfileItem(context, profile);
+        orcidClient.revokeToken(profileToken);
+    }
+
+    private void clearOrcidProfileMetadata(Context context, Item profile) throws SQLException {
+        itemService.clearMetadata(context, profile, "person", "identifier", "orcid", Item.ANY);
+        itemService.clearMetadata(context, profile, "dspace", "orcid", "scope", Item.ANY);
+        itemService.clearMetadata(context, profile, "dspace", "orcid", "authenticated", Item.ANY);
     }
 
     @Override
@@ -249,20 +254,6 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
         return OrcidProfileDisconnectionMode.fromString(value);
     }
 
-    @Override
-    public Iterator<Item> findProfilesByOrcid(Context context, String orcid) {
-        DiscoverQuery discoverQuery = new DiscoverQuery();
-        discoverQuery.setDSpaceObjectFilter(IndexableItem.TYPE);
-        discoverQuery.addFilterQueries("search.entitytype:" + researcherProfileService.getProfileType());
-        discoverQuery.addFilterQueries("person.identifier.orcid:" + orcid);
-        return new DiscoverResultItemIterator(context, discoverQuery);
-    }
-
-    @Override
-    public Iterator<Item> findProfilesWithOrcid(Context context) {
-        return findProfilesByOrcid(context, "*");
-    }
-
     private void setAccessToken(Context context, Item profile, EPerson ePerson, String accessToken) {
         OrcidToken orcidToken = orcidTokenService.findByEPerson(context, ePerson);
         if (orcidToken == null) {
@@ -302,6 +293,11 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
 
     private void clearSynchronizationSettings(Context context, Item profile)
         throws SQLException {
+
+        if (configurationService.getBooleanProperty("orcid.disconnection.remain-sync", false)) {
+            return;
+        }
+
         itemService.clearMetadata(context, profile, "dspace", "orcid", "sync-mode", Item.ANY);
         itemService.clearMetadata(context, profile, "dspace", "orcid", "sync-profile", Item.ANY);
 
@@ -355,6 +351,21 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
             ePersonService.update(context, ePerson);
         } catch (AuthorizeException e) {
             throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public List<Item> findProfilesByOrcid(Context context, String orcid) {
+        DiscoverQuery discoverQuery = new DiscoverQuery();
+        discoverQuery.setDSpaceObjectFilter(IndexableItem.TYPE);
+        discoverQuery.addFilterQueries("search.entitytype:" + researcherProfileService.getProfileType());
+        discoverQuery.addFilterQueries("person.identifier.orcid:" + orcid);
+        try {
+            return searchService.search(context, discoverQuery).getIndexableObjects().stream()
+                .map(object -> ((IndexableItem) object).getIndexedObject())
+                .collect(Collectors.toList());
+        } catch (SearchServiceException ex) {
+            throw new RuntimeException(ex);
         }
     }
 }
