+++ dspace-api/src/main/java/org/dspace/scripts/ProcessServiceImpl.java
@@ -19,11 +19,8 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Pattern;
@@ -48,8 +45,8 @@ import org.dspace.core.Context;
 import org.dspace.core.LogHelper;
 import org.dspace.eperson.EPerson;
 import org.dspace.eperson.Group;
-import org.dspace.eperson.service.GroupService;
 import org.dspace.scripts.service.ProcessService;
+import org.dspace.services.ConfigurationService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -63,9 +60,6 @@ public class ProcessServiceImpl implements ProcessService {
     private ProcessDAO processDAO;
 
     @Autowired
-    private GroupService groupService;
-
-    @Autowired
     private BitstreamService bitstreamService;
 
     @Autowired
@@ -76,6 +70,9 @@ public class ProcessServiceImpl implements ProcessService {
 
     @Autowired
     private MetadataFieldService metadataFieldService;
+
+    @Autowired
+    private ConfigurationService configurationService;
 
     @Override
     public Process create(Context context, EPerson ePerson, String scriptName,
@@ -95,11 +92,17 @@ public class ProcessServiceImpl implements ProcessService {
             });
 
         Process createdProcess = processDAO.create(context, process);
-        String message = Objects.nonNull(ePerson) ? "eperson with email " + ePerson.getEmail() : "an Anonymous user";
-        log.info(LogHelper.getHeader(context, "process_create",
-                                      "Process has been created for " + message
-                                          + " with process ID " + createdProcess.getID() + " and scriptName " +
-                                          scriptName + " and parameters " + parameters));
+
+        if (ePerson != null) {
+            log.info(LogHelper.getHeader(context, "process_create",
+                "Process has been created for eperson with email " + ePerson.getEmail()
+                    + " with ID " + createdProcess.getID() + " and scriptName " +
+                    scriptName + " and parameters " + parameters));
+        } else {
+            log.info(LogHelper.getHeader(context, "process_create",
+                "Process has been created for command-line user with ID " + createdProcess.getID()
+                    + " and scriptName " + scriptName + " and parameters " + parameters));
+        }
         return createdProcess;
     }
 
@@ -171,25 +174,6 @@ public class ProcessServiceImpl implements ProcessService {
     @Override
     public void appendFile(Context context, Process process, InputStream is, String type, String fileName)
         throws IOException, SQLException, AuthorizeException {
-        Bitstream bitstream = createFileBitstream(context, process, is, type, fileName);
-        Map<Integer, Group> groupPolicy = null;
-        Map<Integer, EPerson> userPolicy = null;
-        if (Objects.isNull(context.getCurrentUser())) {
-            Group anonymous = groupService.findByName(context, Group.ANONYMOUS);
-            groupPolicy = new HashMap<>();
-            groupPolicy.put(Constants.READ, anonymous);
-        } else {
-            userPolicy = new HashMap<>();
-            userPolicy.put(Constants.READ, context.getCurrentUser());
-            userPolicy.put(Constants.WRITE, context.getCurrentUser());
-            userPolicy.put(Constants.DELETE, context.getCurrentUser());
-        }
-        this.addBitstream(context, process, bitstream, type, groupPolicy, userPolicy);
-    }
-
-    private Bitstream createFileBitstream(Context context, Process process, InputStream is, String type,
-            String fileName)
-            throws IOException, SQLException {
         Bitstream bitstream = bitstreamService.create(context, is);
         if (getBitstream(context, process, type) != null) {
             throw new IllegalArgumentException("Cannot create another file of type: " + type + " for this process" +
@@ -200,41 +184,10 @@ public class ProcessServiceImpl implements ProcessService {
         MetadataField dspaceProcessFileTypeField = metadataFieldService
             .findByString(context, Process.BITSTREAM_TYPE_METADATAFIELD, '.');
         bitstreamService.addMetadata(context, bitstream, dspaceProcessFileTypeField, null, type);
-        return bitstream;
-    }
-
-    @Override
-    public void appendFile(Context context, Process process, InputStream is, String type, String fileName,
-            Map<Integer, Group> groupPolicy, Map<Integer, EPerson> userPolicy)
-            throws IOException, SQLException, AuthorizeException {
-        this.addBitstream(context, process, createFileBitstream(context, process, is, type, fileName), type,
-                groupPolicy, userPolicy);
-    }
-
-    private void addBitstream(Context context, Process process, Bitstream bitstream, String type,
-            Map<Integer, Group> groupPolicy, Map<Integer, EPerson> userPolicy)
-            throws IOException, SQLException, AuthorizeException {
-
-        if (Objects.nonNull(groupPolicy)) {
-            for (Map.Entry<Integer, Group> entry : groupPolicy.entrySet()) {
-                this.authorizeService.addPolicy(context, bitstream, entry.getKey(), entry.getValue());
-            }
-        }
-
-        if (Objects.nonNull(userPolicy)) {
-            for (Map.Entry<Integer, EPerson> entry : userPolicy.entrySet()) {
-                this.authorizeService.addPolicy(context, bitstream, entry.getKey(), entry.getValue());
-            }
-        }
-
-        try {
-            context.turnOffAuthorisationSystem();
-            bitstreamService.update(context, bitstream);
-            context.restoreAuthSystemState();
-        } catch (SQLException | AuthorizeException e) {
-            log.info(e.getMessage());
-            throw new RuntimeException(e.getMessage(), e);
-        }
+        authorizeService.addPolicy(context, bitstream, Constants.READ, context.getCurrentUser());
+        authorizeService.addPolicy(context, bitstream, Constants.WRITE, context.getCurrentUser());
+        authorizeService.addPolicy(context, bitstream, Constants.DELETE, context.getCurrentUser());
+        bitstreamService.update(context, bitstream);
         process.addBitstream(bitstream);
         update(context, process);
     }
@@ -336,16 +289,12 @@ public class ProcessServiceImpl implements ProcessService {
         return processDAO.countTotalWithParameters(context, processQueryParameterContainer);
     }
 
-    @Override
-    public int countByUser(Context context, EPerson user) throws SQLException {
-        return processDAO.countByUser(context, user);
-    }
 
     @Override
     public void appendLog(int processId, String scriptName, String output, ProcessLogLevel processLogLevel)
             throws IOException {
-        File tmpDir = FileUtils.getTempDirectory();
-        File tempFile = new File(tmpDir, scriptName + processId + ".log");
+        File logsDir = getLogsDirectory();
+        File tempFile = new File(logsDir, processId + "-" + scriptName + ".log");
         FileWriter out = new FileWriter(tempFile, true);
         try {
             try (BufferedWriter writer = new BufferedWriter(out)) {
@@ -360,18 +309,43 @@ public class ProcessServiceImpl implements ProcessService {
     @Override
     public void createLogBitstream(Context context, Process process)
             throws IOException, SQLException, AuthorizeException {
-        File tmpDir = FileUtils.getTempDirectory();
-        File tempFile = new File(tmpDir, process.getName() + process.getID() + ".log");
-        FileInputStream inputStream = FileUtils.openInputStream(tempFile);
-        appendFile(context, process, inputStream, Process.OUTPUT_TYPE, process.getName() + process.getID() + ".log");
-        inputStream.close();
-        tempFile.delete();
+        File logsDir = getLogsDirectory();
+        File tempFile = new File(logsDir, process.getID() + "-" + process.getName() + ".log");
+        if (tempFile.exists()) {
+            FileInputStream inputStream = FileUtils.openInputStream(tempFile);
+            appendFile(context, process, inputStream, Process.OUTPUT_TYPE,
+                       process.getID() + "-" + process.getName() + ".log");
+            inputStream.close();
+            tempFile.delete();
+        }
     }
 
     @Override
     public List<Process> findByStatusAndCreationTimeOlderThan(Context context, List<ProcessStatus> statuses,
         Date date) throws SQLException {
         return this.processDAO.findByStatusAndCreationTimeOlderThan(context, statuses, date);
+    }
+
+    @Override
+    public int countByUser(Context context, EPerson user) throws SQLException {
+        return processDAO.countByUser(context, user);
+    }
+
+    @Override
+    public void failRunningProcesses(Context context) throws SQLException, IOException, AuthorizeException {
+        List<Process> processesToBeFailed = findByStatusAndCreationTimeOlderThan(
+                context, List.of(ProcessStatus.RUNNING, ProcessStatus.SCHEDULED), new Date());
+        for (Process process : processesToBeFailed) {
+            context.setCurrentUser(process.getEPerson());
+            // Fail the process.
+            log.info("Process with ID {} did not complete before tomcat shutdown, failing it now.", process.getID());
+            fail(context, process);
+            // But still attach its log to the process.
+            appendLog(process.getID(), process.getName(),
+                      "Process did not complete before tomcat shutdown.",
+                      ProcessLogLevel.ERROR);
+            createLogBitstream(context, process);
+        }
     }
 
     private String formatLogLine(int processId, String scriptName, String output, ProcessLogLevel processLogLevel) {
@@ -389,4 +363,15 @@ public class ProcessServiceImpl implements ProcessService {
         return  sb.toString();
     }
 
+    private File getLogsDirectory() {
+        String pathStr = configurationService.getProperty("dspace.dir")
+            + File.separator + "log" + File.separator + "processes";
+        File logsDir = new File(pathStr);
+        if (!logsDir.exists()) {
+            if (!logsDir.mkdirs()) {
+                throw new RuntimeException("Couldn't create [dspace.dir]/log/processes/ directory.");
+            }
+        }
+        return logsDir;
+    }
 }
